Fix WFS-T to use correct url, and also fix subsequent requests in same server (by using expunge_all)
